перенесены компоненты в папку shared
