+++ src/app/App.js
@@ -4,9 +4,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import React, { useEffect } from 'react';
 
 // Локальные импорты из проекта
+import { ButtonScrollUp, ScrollToTop } from '../shared/ui/ButtonScrollUp';
 import Routing from '../pages';
 import OrdersList from '../entities/orders/ui/OrdersList/OrdersList';
-import ButtonScrollUp from '../components/UI/ButtonScrollUp/ButtonScrollUp';
 import OrderPopup from '../components/PopupsRedux/OrderPopup/OrderPopup';
 import Header from '../components/Header/Header';
 import Footer from '../components/Footer/Footer';
@@ -15,6 +15,7 @@ import AuthPopup from '../components/AuthPopup/AuthButtons/AuthPopup';
 import { getUser } from './store/slices/accountSlice';
 
 import './App.scss';
+
 
 // Компонент app
 function App() {
@@ -28,17 +29,6 @@ function App() {
       dispatch(getUser(auth_token));
     }
   }, [isFetched, isLoggedIn]);
-
-  // Компонент ScrollToTop для прокрутки страницы наверх при смене маршрута
-  const ScrollToTop = () => {
-    const location = useLocation();
-
-    useEffect(() => {
-      window.scrollTo(0, 0);
-    }, [location.pathname]);
-
-    return null;
-  };
 
   return (
     <div className="app">
