+++ src/shared/ui/ButtonScrollUp/index.js
@@ -1,0 +1,2 @@
+export { ButtonScrollUp } from './ButtonScrollUp';
+export { ScrollToTop } from './ScrollToTop';
