+++ src/shared/ui/Counter/Counter.scss
