+++ src/shared/ui/Layout/Main/Main.scss
@@ -3,6 +3,6 @@
 .main {
   @include box-container;
   @include flex-container;
-  margin: 60px 0 0;
+  margin: 60px 0 120px;
   width: 100%;
-}
+  }
