+++ src/shared/ui/RadioButton/index.js
@@ -1,0 +1,1 @@
+export { RadioButton } from './RadioButton';
